+++ rp2040-hal/src/xosc.rs
@@ -49,11 +49,36 @@ pub enum Error {
 }
 
 /// Blocking helper method to setup the XOSC without going through all the steps.
+///
+/// This uses a startup_delay_multiplier of 64, which is a rather conservative value
+/// that should work even if the XOSC starts up slowly. In case you need a fast boot
+/// sequence, and your XOSC starts up quickly enough, use [`setup_xosc_blocking_custom_delay`].
 pub fn setup_xosc_blocking(
     xosc_dev: XOSC,
     frequency: HertzU32,
 ) -> Result<CrystalOscillator<Stable>, Error> {
-    let initialized_xosc = CrystalOscillator::new(xosc_dev).initialize(frequency)?;
+    let initialized_xosc = CrystalOscillator::new(xosc_dev).initialize(frequency, 64)?;
+
+    let stable_xosc_token = nb::block!(initialized_xosc.await_stabilization()).unwrap();
+
+    Ok(initialized_xosc.get_stable(stable_xosc_token))
+}
+
+/// Blocking helper method to setup the XOSC without going through all the steps.
+///
+/// This function allows setting a startup_delay_multiplier to tune the amount of time
+/// the chips waits for the XOSC to stabilize.
+/// The default value in the C SDK is 1, which should work on the Raspberry Pico, and many
+/// third-party boards.
+/// [`setup_xosc_blocking`], uses a conservative value of 64, which is the value commonly
+/// used on slower-starting oscillators.
+pub fn setup_xosc_blocking_custom_delay(
+    xosc_dev: XOSC,
+    frequency: HertzU32,
+    startup_delay_multiplier: u32,
+) -> Result<CrystalOscillator<Stable>, Error> {
+    let initialized_xosc =
+        CrystalOscillator::new(xosc_dev).initialize(frequency, startup_delay_multiplier)?;
 
     let stable_xosc_token = nb::block!(initialized_xosc.await_stabilization()).unwrap();
 
@@ -91,7 +116,12 @@ impl CrystalOscillator<Disabled> {
     }
 
     /// Initializes the XOSC : frequency range is set, startup delay is calculated and set.
-    pub fn initialize(self, frequency: HertzU32) -> Result<CrystalOscillator<Initialized>, Error> {
+    /// Set startup_delay_multiplier to a value > 1 when using a slow-starting oscillator.
+    pub fn initialize(
+        self,
+        frequency: HertzU32,
+        startup_delay_multiplier: u32,
+    ) -> Result<CrystalOscillator<Initialized>, Error> {
         const ALLOWED_FREQUENCY_RANGE: RangeInclusive<HertzU32> =
             HertzU32::MHz(1)..=HertzU32::MHz(15);
         //1 ms = 10e-3 sec and Freq = 1/T where T is in seconds so 1ms converts to 1000Hz
@@ -100,6 +130,10 @@ impl CrystalOscillator<Disabled> {
 
         if !ALLOWED_FREQUENCY_RANGE.contains(&frequency) {
             return Err(Error::FrequencyOutOfRange);
+        }
+
+        if startup_delay_multiplier == 0 {
+            return Err(Error::BadArgument);
         }
 
         self.device.ctrl.write(|w| {
@@ -112,6 +146,7 @@ impl CrystalOscillator<Disabled> {
         //See Chapter 2, Section 16, §3)
         //We do the calculation first.
         let startup_delay = frequency.to_Hz() / (STABLE_DELAY_AS_HZ.to_Hz() * DIVIDER);
+        let startup_delay = startup_delay.saturating_mul(startup_delay_multiplier);
 
         //Then we check if it fits into an u16.
         let startup_delay: u16 = startup_delay.try_into().map_err(|_| Error::BadArgument)?;